bedrockagentcore/requirements.txt
strands-agents
strands-agents-tools
strands-agents[litellm]
strands-agents[otel]
uv
boto3
langchain[aws]
langgraph
langsmith[otel]
duckduckgo-search
langchain-community
aws-opentelemetry-distro
opentelemetry-instrumentation-langchain
bedrock-agentcore
bedrock-agentcore-starter-toolkit
pydantic
httpx
ipython
ipykernel
pandas
jupyterlab
openpyxl
mcp>=1.9.0
fastapi 
uvicorn[standard]